# Improve performance of queries that return lots of nodes