@OskarDamkjaer 2021-12-14
Lovely to see such a nice performance improvement 😍 I was wondering if there was some neat way we could make sure we don't regress again on this function... I went looking for some good way to add a "benchmarking" unit test, but the problem is that it depends too much on outside factors (like how fast your laptop happens to be). 

For the future I think it'd be interesting if we leave some ways in the code so that manual testing of performance is easier something like the code checking if `node.ENV === "TEST"` and then running `performance.mark`. 🤔